# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's Xlearn",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://servicios.10xconsultores.org"
        },
        {
            "url": "http://127.0.0.1:8000"
        },
        {
            "url": "https://servicios.asstiseguridadsocial.com"
        }
    ],
    "paths": {
...
        "/api/v1/diagnostic/store": {
            "post": {
                "tags": [
                    "Diagnostic"
                ],
                "summary": "Almacenar diagnostico del lider",
                "operationId": "c624d5dc0cdfedb7d92f4cc386caa5b5",
                "parameters": [
                    {
                        "name": "target",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "_rel",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "answers",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "[]"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Registro almacenado con éxito.",
                                        "course_route": "[]"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/diagnostic/confirm_route/{diagnostic_id}": {
            "patch": {
                "tags": [
                    "Diagnostic"
                ],
                "summary": "Confirmar diagnostico",
                "operationId": "1a573a8890a77720c6eca49c03eabc0d",
                "parameters": [
                    {
                        "name": "diagnostic_id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Ruta confirmada correctamente"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}